Automatically tag gem builds as Date.today
When you build a gem, the gemspec is evaluated and re-serialized. All
dynamic values become static, once again. (for evidence of this, compare
ruby-progressbar.gemspec with the installed gemspec in your rubygems
folder.  Also, if you untar a gem with xzf, and examine metadata.gz,
you'll see the gemspec is actually stored in serialized YAML format, and
then converted back into a ruby format).

Therefore, this is a safe way to automatically store the current date in
a gem when you invoke gem build.  Currently, it's stuck on a Feb 2009
date, as evidence by the rubyforge page:

http://rubygems.org/gems/ruby-progressbar