Feature: The running average is always set back to 0 when the bar is reset
Previously when the bar was reset or restarted, the running average would not
also be reset.  Additionally if the bar was started at some arbitrary point in
the middle of progress, it should not have a running average at that point.

This change (by referencing progress made) rather than total progress, fixes
both issues.

--------------------------------------------------------------------------------
Github Issue URLs:
  * https://github.com/jfelchner/ruby-progressbar/issues/66
Actions:
  * Closes #66
Branch:        feature/fix-running-average-when-bar-is-reset
Change-Id:     I211e53846faa784b78667c956f449148539ca458
Signed-off-by: Jeff Felchner <accounts+git@thekompanee.com>