Refactor: Change Length Calculator from Module to Class
We wanted Length Calculator to be an object rather than a module that we just
keep adding methods to, in order to decouple it from the progress bar itself.

We made that basic change, and then we changed all the references to an instance
variable for the new class we made in order to make all the tests pass.

The changes in the tests are just hacks in order to get them to pass. We will
need to go back later and fix it in a more elegant way.

--------------------------------------------------------------------------------
Branch:        refactor/massive-update
Signed-off-by: Jeff Felchner <accounts+git@thekompanee.com>