Refactor: Extract Timer out into its own class
There will now be one 'master' timer for the entire bar and that timer will be
passed into the different components which require it.

Unlike previously, the components will not be responsible for updating the
timer, but will instead merely reference it.

This will eliminate the need for cycling through all of the components simply to
send them the same message.

--------------------------------------------------------------------------------
Branch:        refactor/massive-update
Signed-off-by: Jeff Felchner <accounts+git@thekompanee.com>