Refactor: Change Progressable from a module into a class
For similar reasons as to why we changed `Timer` into a class, there's only need
to have one object tracking the progress of the bar.

Instead of mixing in the functionality for tracking progress into multiple
components, instead those components will be passed the `Progressable` and ask
it for information.

Also, similarly to the `Timer`, we removed all code which had the components
manipulating the progress and instead all progress manipulation happens in the
`Base`.

--------------------------------------------------------------------------------
Branch:        refactor/massive-update
Signed-off-by: Jeff Felchner <accounts+git@thekompanee.com>