Refactor: Remove ElapsedTimer component
Now that the `Timer` is extracted into its own class, the `ElapsedTime`
component was extremely thin.  Looking at it, and how we were using it as well
as the `EstimatedTime` component, made it obvious that the two should be
combined.

This is the first step.  The next couple commits will involve renaming.

--------------------------------------------------------------------------------
Branch:        refactor/massive-update
Signed-off-by: Jeff Felchner <accounts+git@thekompanee.com>