Refactor: Move 'autofinish' out into 'Base'
Another awful commit (last one I promise).

Moved `autofinish` out into the `Base` because `Progress` should only care about
whether the `progress` is equal to the `total` to tell whether or not it's
'finished'.

The `Base` is the thing that should be responsible for coordinating between
`progress` and `timer` to get the answer as to whether it's actually 'finished'.

Additionally I changed some unneeded references of `self.some_var` to just
`some_var` and inlined a couple methods.

--------------------------------------------------------------------------------
Branch:        refactor/massive-update
Signed-off-by: Jeff Felchner <accounts+git@thekompanee.com>