Refactor: Extract 'TtyOutput' and 'NonTtyOutput'
This allows us to get rid of all of the nasty 'if' statements in the `Output`
class.  When we create the output in `Base`, it detects what type of output the
user would like.

From that point forward, we can just call methods on it and, since it's of the
proper type, we know that it's going to either do TTY or non-TTY logic.

This also opens up the possibility of adding output types like 'colored', etc.

--------------------------------------------------------------------------------
Branch:        refactor/massive-update
Signed-off-by: Jeff Felchner <accounts+git@thekompanee.com>