Refactor: Create a Title component
This was really the last instance of a thing in `Base` that needed to be
converted to a component.  I wasn't going to do it unless necessary since it was
so small, but since we want to make all the molecules access the same way, it
ended up being necessary and at the same time eliminates a special case.

--------------------------------------------------------------------------------
Branch:        refactor/massive-update
Signed-off-by: Jeff Felchner <accounts+git@thekompanee.com>