Refactor: Performance tuning
A couple of changes which keep things from being executed more times than they
need to be.

In the case of the throttle change, my assumption is that we don't need to clear
the bar if we're not going to update the bar.  If we're not going to clear the
bar, then we don't need to see if the length changed.

In the case of the time mock, once we know that time is being mocked by a given
method, we don't need to check it every single time.  `respond_to?` is
expensive.

--------------------------------------------------------------------------------
Branch:        refactor/massive-update
Signed-off-by: Jeff Felchner <accounts+git@thekompanee.com>