Fixed: Progress#finish causing an exception when total was unknown
If `Progress#finish` was called when the total was unknown, it would raise an
exception.

The base issue is that `progress` was getting set to `nil` since `total` was
`nil`.  When we call `finish` and there's no `total`, then `finish` is basically
a noop.

Since `progress` will never be `nil`, then any time `total` _is_ `nil` and
`finished?` is called, the answer will be `false`, which is correct because
a bar with an unknown total is by default never finished.

Fixes: #101