Add: Ability to specify a completely custom output stream
If you pass a class in as the `output` option which is a subclass of
`ProgressBar::Output`, it will be instantiated with all of the options
passed to any of the other built-in output classes.

Using this method, a user can completely customize how the bar is sent
to an IO stream.

This change is not going to be considered public API and _no attempt_
will be made to follow SemVer.

USE AT YOUR OWN RISK!

------------------------------------------------------------------------
Change-Id:     Ie8f73c06143098cf26621ba03e9ca5b6056585c9