Performance: don't shell out when it's avoidable.
When running in Zeus, `IO.console` was returning `nil` so
ruby-progressbar was resorting to shelling out to `stty` and `tput` to
get the console width.  Launching a process on every increment adds a
lot of overhead time to the progress bar. (1/20 of a second on my Linux
machine.)

In this fix, I pass the output stream to the length calculator. If the
stream is a tty, I call its `winsize` method instead of launching
processes.