Remove: allowed_push_host From gemspecs
Why This Change Is Necessary
========================================================================

We want to allow users to be able to push to other internal hosts which
may not be rubygems.org.

How These Changes Address the Issue
========================================================================

Removes the `allowed_push_host` configuration from the gemspec.

Side Effects Caused By This Change
========================================================================

The gem can now be pushed to any server URL.

------------------------------------------------------------------------
Actions:
  * Closes #151