Refactor: Simplify Enumerator Refinement
Why This Change Is Necessary
========================================================================

The [original PR][1] was submitted by @ronen and at the time I had not
looked into it sufficiently.  It seems as though this code can be
greatly simplified by leveraging the fact that we are already refining
an `Enumerator` and therefore there is no need to wrap in another
`Enumerator`.

What These Changes Do To Address the Issue
========================================================================

Remove the wrapping `Enumerator`

Caveats
------------------------------------------------------------------------

I'd love to have @ronen take a look at this and see if anything jumps
out as being very wrong about this implementation.

Side Effects Caused By This Change
========================================================================

None expected.

------------------------------------------------------------------------

**Relevant URLs:**

  * [Original PR Link][1]

------------------------------------------------------------------------

[1]: https://github.com/jfelchner/ruby-progressbar/pull/126