Add: Ability To Access Progress Bar Instance Using Enumerator Refinement
Why This Change Is Necessary
========================================================================

Without access to the progress bar instance, when using the refinement
there's a lack of ability to do things like log or modify the title
mid-progress like you can with the standard progress bar usage.

What These Changes Do To Address the Issue
========================================================================

If you pass a second argument to the block passed to `with_progressbar`,
the second argument will be the progress bar instance.

Side Effects Caused By This Change
========================================================================

None expected.

------------------------------------------------------------------------

**Relevant URLs:**

  * [Current Ticket Link][1]
  * [Original Ticket Link][2]

**Actions:**

  * Closes 184
  * Closes 185

------------------------------------------------------------------------

[1]: https://github.com/jfelchner/ruby-progressbar/issues/185
[2]: https://github.com/jfelchner/ruby-progressbar/issues/184