Refactor: Change Refinement To Understand That each Passes One Item
Why This Change Is Necessary
========================================================================

`each` does not deal with multiple items, it always yields [a single
element][2] and so we can refrain from using `args` there.

What These Changes Do To Address the Issue
========================================================================

Switch from `*args` to `item`.

Caveats
------------------------------------------------------------------------

We are still going to pass `*yielded_args` to the `yield` because in the
next commit we will pass a second argument to that `yield` call.

Side Effects Caused By This Change
========================================================================

None expected.

------------------------------------------------------------------------

**Relevant URLs:**

  * [Ticket Link][1]
  * [Ruby Doc For `each`][2]

------------------------------------------------------------------------

[1]: https://github.com/jfelchner/ruby-progressbar/issues/185
[2]: https://ruby-doc.org/2.7.6/Enumerator.html#method-i-each