Change: Allow Timer's Elapsed Seconds To Calculate Even If Not Started
Why This Change Is Necessary
========================================================================

Because the typical use case is for the progress bar to be started
automatically, this didn't typically surface as a bug, however in [this
issue][1] it was brought up that when the bar is reset (which causes it
to remove its start time) there is an issue with trying to query its
`elapsed_seconds` (which _requires_ a start time).

What These Changes Do To Address the Issue
========================================================================

Modify `elapsed_seconds` so that it returns a correct value even if it
is not started.

Side Effects Caused By This Change
========================================================================

None expected.

------------------------------------------------------------------------

**Relevant URLs:**

  * [Ticket Link][1]

**Actions:**

  * Closes #177

------------------------------------------------------------------------

[1]: https://github.com/jfelchner/ruby-progressbar/issues/177