Add: Wall Clock Time To Completion To The Format Elements
Why This Change Is Necessary
========================================================================

It has [been requested][1] to add the wall clock time of completion to
ease the computation necessary to determine when the progress will be
complete.

What These Changes Do To Address the Issue
========================================================================

Adds `%l` as a new format element to display the wall clock time.

Caveats
------------------------------------------------------------------------

Because this is a convenience item it is designed to be used locally and
therefore the wall clock time is displayed in whatever the local time of
the user's system is.

Side Effects Caused By This Change
========================================================================

None expected.

------------------------------------------------------------------------

**Relevant URLs:**

  * [Ticket Link][1]

**Actions:**

  * Closes #161

------------------------------------------------------------------------

[1]: https://github.com/jfelchner/ruby-progressbar/issues/161#issuecomment-641070396