Add: Concept Of A Projector
Why This Change Is Necessary
========================================================================

Currently the `Progress` class is doing too much.  It's both in charge
of keeping track of the progress but also in understanding how quickly
each item of progress is being executed.

We need to split the objects up so that each object is responsible for
its own job.

What These Changes Do To Address the Issue
========================================================================

Introduce the concept of a "projector".

Caveats
------------------------------------------------------------------------

Currently this is simply the calculator but will be fleshed out in later
commits.

Side Effects Caused By This Change
========================================================================

None expected.