Remove: smoothing Option
Why This Change Is Necessary
========================================================================

The plan is to allow different types of projectors to be used by the gem
at some point and not all of those projectors are going to need the
option to smooth, so instead we will change the configuration to allow
for each projector to have their own configuration.

What These Changes Do To Address the Issue
========================================================================

We add a deprecation warning for anyone using `smoothing` as an option
and also add the new approach.

The new approach will be instead of passing this:

```ruby
ProgressBar.create(smoothing: 0.1)
```

You would pass:

```ruby
ProgressBar.create(projector: { smoothing: 0.1 })
```

Side Effects Caused By This Change
========================================================================

Additional warnings for anyone using `smoothing`.