Refactor: Update Projection To Keep Track Of Past Projections
Why This Change Is Necessary
========================================================================

Previous projections are valuable information for some projectors and so
each projector needs to have the capability of tracking it if it so
chooses.

What These Changes Do To Address the Issue
========================================================================

Track the projection in a new variable.

Side Effects Caused By This Change
========================================================================

None expected.