Refactor: Send All Progress Events To The Projector
Why This Change Is Necessary
========================================================================

The projectors need to know when progress changes so that eventually it
will be able to update the projection each time it does.

What These Changes Do To Address the Issue
========================================================================

Start calling the methods on the projector but stub out the various
progress methods as noops.

Side Effects Caused By This Change
========================================================================

None expected.