Refactor: Pass Projector To The Time Component
Why This Change Is Necessary
========================================================================

The time component is the primary use case for the projector since it
needs to represent the estimated amount of time left in the run.

Currently the time component is getting the projection off the
`Progress` object but at this point it is able to go directly to the
`Projector`.

What These Changes Do To Address the Issue
========================================================================

Pass the projector to the time component and use it instead of the
progress.

Side Effects Caused By This Change
========================================================================

None expected.