Refactor: Projector To Be Able To Keep Track Of Progress
Why This Change Is Necessary
========================================================================

Previously we were passing new values into `calculate` but were not
saving them.

In order to be able to be more useful in more situations, we need to
have the projector be called each time progress is updated and be able
to save that data if necessary for future projections.

This particular calculator only needs two values (the starting point and
the current point) but other projectors may need more (for example
a projector that only cares about the last n minutes of items or one
that only cares about the last x items).

What These Changes Do To Address the Issue
========================================================================

Because the projector is now tied into progress events, all we have to
do is implement those methods and save the data as necessary.

We introduce a concept of "samples" into the projector.  Each projector
will be responsible for holding as many samples of progress data as is
necessary for it to do its job properly.

Side Effects Caused By This Change
========================================================================

None expected.