Refactor: The Progress No Longer Needs The Projector
Why This Change Is Necessary
========================================================================

We've removed all usages of the projector from the `Progress` object and
therefore it is no longer needed to do its job.

What These Changes Do To Address the Issue
========================================================================

Remove all usages of the projector from `Progress` and update test setup
to reflect that.

Side Effects Caused By This Change
========================================================================

None expected.