Add: Ability To Allow Different Projectors To Be Used For Time Estimates
Why This Change Is Necessary
========================================================================

The current projector we've had in the code for over a decade now has
served us well, but there has been a desire to have some different types
of projectors (eg windowed by item count or by time).  This was
previously difficult to do but with this change we should be able to
allow the user to swap different ones out fairly easily.

What These Changes Do To Address the Issue
========================================================================

Add a new `Projector` class that manages all other projectors.

Side Effects Caused By This Change
========================================================================

None expected.