Refactor: Rename Calculators To Projectors
Why This Change Is Necessary
========================================================================

Now that the calculator is responsible for more than just a single
`calculate` method, it is being renamed to something more appropriate
for its task.

What These Changes Do To Address the Issue
========================================================================

Rename and move the files for the `SmoothedAverage` class.

Side Effects Caused By This Change
========================================================================

None expected.