Let ruby-progressbar and progressbar gems allow requiring progressbar.rb or ruby-progessbar.rb
Having the 2 gems with quite the same content but not exactly is very
confusing, for instance on Debian because the ruby-progressbar deb
package actually contains the progressbar gem, not the ruby-progressbar
gem.

IMHO, having both gems with the exact same content would be more
user-friendly, allowing 'require "progressbar"' and 'require
"ruby-progressbar"' indifferently from what gem is actually installed
(be it with gem install or a distrubution package).